pyproject.toml
[tool.poetry]
name = "python-flask-basic"
version = "0.1.0"
description = ""
authors = ["euiyounghwang <marieuig@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.4.4"
python-dotenv = "^1.0.1"
flask = "^3.0.1"
connexion = {extras = ["swagger-ui"], version = "^3.0.5"}
pytz = "^2023.3.post1"
flask-cors = "^4.0.0"
gunicorn = "^21.2.0"
uvicorn = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
